Add trait and interface for base employee
